fix tests and break up llm edge function
