+++ supabase/tests/functions/process-llm-query.integration.test.ts
@@ -1,10 +1,14 @@
+// supabase/tests/functions/process-llm-query.integration.test.ts
+
 import { assertEquals } from "https://deno.land/std@0.207.0/assert/mod.ts";
-import { createClient } from '@supabase/supabase-js';
-import { handleRequest } from "../../functions/process-llm-query/index.ts";
+import { createClient } from 'jsr:@supabase/supabase-js@2';
+import { handleRequest } from "../../functions/process-llm-query/handler.ts";
+import type { Dependencies } from "../../functions/process-llm-query/types.ts";
 
 Deno.test("integration - successfully processes query", async () => {
   // Mock OpenAI response
-  const mockFetch = async (url: string, init?: RequestInit) => {
+  const mockFetch: Dependencies['fetch'] = async (input, init?) => {
+    const url = input.toString();
     if (url.includes('openai')) {
       return new Response(JSON.stringify({
         choices: [{
@@ -38,36 +42,14 @@ Deno.test("integration - successfully processes query", async () => {
     })
   });
 
-  const response = await handleRequest(
-    mockRequest,
-    { OPENAI_API_KEY: "test_key" },
-    {
-      supabaseClient: mockSupabase as unknown as ReturnType<typeof createClient>,
-      fetchFn: mockFetch
-    }
-  );
+  const response = await handleRequest(mockRequest, {
+    supabaseClient: mockSupabase as unknown as ReturnType<typeof createClient>,
+    fetch: mockFetch,
+    openaiApiKey: "test_key"
+  });
 
   const data = await response.json();
   assertEquals(response.status, 200);
   assertEquals(data.success, true);
   assertEquals(data.responses_count, 1);
-});
-
-Deno.test("integration - handles missing OpenAI key", async () => {
-  const mockRequest = new Request("http://localhost", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      query_id: "test-id",
-      prompt: "test question",
-      response_types: ["open_ended"],
-      hive_size: 1
-    })
-  });
-
-  const response = await handleRequest(mockRequest, {});
-  const data = await response.json();
-  
-  assertEquals(response.status, 500);
-  assertEquals(data.error, "OPENAI_API_KEY not set");
 });
